.github/scripts/requirements.txt
pyaml
jsonschema
